bug fixes with Id

+++ src/resources/tasks/task.memory.repository.js
@@ -29,10 +29,32 @@ const deleteById = async (id) => {
   tasks = tasks.filter((task) => task.id !== id);
 };
 
-const updateById = async (id, title, columns) => {
-  tasks = tasks.map((task) => task.id === id ? {id, title, columns} : task);
+const updateById = async (
+  id, 
+  title,
+  order,
+  description,
+  userId,
+  boardId,
+  columnId) => {
+  tasks = tasks.map((task) => task.id === id ? {
+    id, 
+    title,
+    order,
+    description,
+    userId,
+    boardId,
+    columnId} : task);
   const updatedTask = await getById(id);
   return updatedTask;
 };
 
-module.exports = { getAll, getById, create, deleteById, updateById };
+const updateUserIdToNull = (userId, nullUserId) => {
+  tasks.forEach((task) => {
+    if(task.userId === userId) {
+      Object.assign(task, {"userId": nullUserId});
+    }
+  })     
+}
+
+module.exports = { getAll, getById, create, deleteById, updateById, updateUserIdToNull };
